# Parquet: Use proto instead of json for events