--- a/tempodb/encoding/vparquet/schema.go
+++ b/tempodb/encoding/vparquet/schema.go
@@ -59,7 +59,7 @@ type Attribute struct {
 
 type EventAttribute struct {
 	Key   string `parquet:",snappy,dict"`
-	Value string `parquet:",snappy"` // Json-encoded data
+	Value string `parquet:",snappy"` // Was json-encoded data, is now proto encoded data
 }
 
 type Event struct {
@@ -324,11 +324,12 @@ func eventToParquet(e *v1_trace.Span_Event) Event {
 	}
 
 	for _, a := range e.Attributes {
-		jsonBytes := &bytes.Buffer{}
-		_ = jsonMarshaler.Marshal(jsonBytes, a.Value)
+		b := make([]byte, a.Value.Size())
+		a.Value.MarshalToSizedBuffer(b)
+
 		ee.Attrs = append(ee.Attrs, EventAttribute{
 			Key:   a.Key,
-			Value: jsonBytes.String(),
+			Value: string(b),
 		})
 	}
 
@@ -396,7 +397,13 @@ func parquetToProtoEvents(parquetEvents []Event) []*v1_trace.Span_Event {
 						Value: &v1.AnyValue{},
 					}
 
-					_ = jsonpb.Unmarshal(bytes.NewBufferString(a.Value), protoAttr.Value)
+					// event attributes are currently encoded as proto, but were previously json.
+					//  this code attempts proto first and, if there was an error, falls back to json
+					err := protoAttr.Value.Unmarshal([]byte(a.Value))
+					if err != nil {
+						_ = jsonpb.Unmarshal(bytes.NewBufferString(a.Value), protoAttr.Value)
+					}
+
 					protoEvent.Attributes = append(protoEvent.Attributes, protoAttr)
 				}
 			}